UserDetail and operation  implementation
UserDetail and operation  implementation

+++ ionic/www/js/controllers/otherroles/organizerController.js
@@ -2,17 +2,22 @@ angular
     .module('starter')
     .controller('organizerController', organizerController);
 
-  organizerController.$inject = ['$scope', '$stateParams', '$state', '$rootScope','userInfoService','$ionicModal'];
+  organizerController.$inject = ['$scope', '$stateParams', '$state', '$rootScope','userInfoService','$ionicModal','userAuthenticationService'];
 
-  function organizerController($scope, $stateParams, $state, $rootScope,userInfoService, $ionicModal) {
+  function organizerController($scope, $stateParams, $state, $rootScope,userInfoService, $ionicModal, userAuthenticationService) {
     var vm = this;
     vm.detailAboutActivity = detailAboutActivity;
     vm.updateActivity = updateActivity;
     vm.saveUpdatedActivityDetail = saveUpdatedActivityDetail;
     vm.deleteActivity = deleteActivity;
+    vm.updateUser = updateUser;
     vm.userName = $rootScope.userDetail.data[0].Name;
+    vm.saveUpdatedUserDetail = saveUpdatedUserDetail;
+    vm.closeModel = closeModel;
+    vm.deleteUser = deleteUser;
+    vm.activityDetail = [];
     showActivity();
-  
+    showUser();
     $scope.sortReverse  = false;
     $scope.sortReverse = false;
     $scope.tabs = [{
@@ -20,7 +25,7 @@ angular
             url: 'addClass.html'
         }, {
             title: 'People',
-            url: 'showClass.html'
+            url: 'people.html'
         }];
 
     $scope.currentTab = 'addClass.html';
@@ -33,71 +38,22 @@ angular
         return tabUrl == $scope.currentTab;
     }
 
-    function detailAboutActivity(activity){
-        console.log('activity',activity);
-       /* if(activity.Phone == "" ||activity.Phone == null ){*/
-            userInfoService.getActivityProjectDetail(activity.Project_url).then(function(projectDetails){
-                console.log('projectDetails',projectDetails);
-                userInfoService.getActivityCoordinatorDetail(activity.Coordinator_url).then(function(coordinatorDetails){
-                console.log(coordinatorDetails);
-                ActivityDetailStructure(activity,coordinatorDetails,projectDetails);
-            },function(error){
-                console.log('error',error);
-            });
-            },function(error){
-                console.log('error',error);
-            })
-           
-        /*}*/
-        $ionicModal.fromTemplateUrl('activityDetail.html', {
-            scope: $scope,
-            animation: 'slide-in-up'
-        }).then(function(modal) {
-            $scope.modal = modal;
-            $scope.modal.show();
-        });
-         
-        $scope.closeModal = function() {
-            $scope.modal.hide();
-        };
- 
-
+    function detailAboutActivity(activity){ 
+        $state.go('app.activityDetail',{'activityDetail':activity},{location: false, inherit: false});
     }
 
-    function ActivityDetailStructure(activity,coordinatorDetails){
-        var _activityDetail = {
-            activity_type_id: activity.Activity_type_id,
-            activity_address: activity.Address,
-            activity_coordinator_url: activity.Coordinator_url,
-            activity_email: activity.Email,
-            activity_end_date: activity.End_date,
-            activity_end_time: activity.End_time,
-            activity_name: activity.Name,
-            activity_phone: activity.Phone,
-            activity_project_url: activity.Project_url,
-            activity_recurrence: activity.Recurrence,
-            activity_sb_Region: activity.SB_Region,
-            activity_start_date: activity.Start_date,
-            activity_start_time: activity.Start_time,
-            activity_URL: activity.URL,
-            activity__url: activity._url,
-            coordinator_Address: coordinatorDetails.data.Address,
-            coordinator_Email: coordinatorDetails.data.Email,
-            coordinator_Name: coordinatorDetails.data.Name,
-            coordinator_Phone: coordinatorDetails.data.Phone,
-            coordinator_Profession: coordinatorDetails.data.Profession,
-            coordinator_Role: coordinatorDetails.data.Role,
-            coordinator_SB_Region: coordinatorDetails.data.SB_Region,
-            coordinator_URL: coordinatorDetails.data.URL,
-            coordinator__url: coordinatorDetails.data._url
-        }
-         vm.activityDetail = _activityDetail;
-       
-    }
-
+    
     function updateActivity(activity){
-        console.log('activity',activity);
+        userInfoService.getAllActivity().then(function(activity){
+            vm.activityList= activity.data ;
+        })
+        userInfoService.getAllProject().then(function(project){
+            vm.projectList = project.data;
+        })
+        
         vm.editActivity = activity;
+
+        console.log('showUp',vm.editActivity);
         $ionicModal.fromTemplateUrl('editActivity.html', {
             scope: $scope,
             animation: 'slide-in-up'
@@ -105,16 +61,15 @@ angular
             $scope.modal = modal;
             $scope.modal.show();
         });
-         
-        $scope.closeModal = function() {
-            $scope.modal.hide();
-        };
  
     }
 
     function saveUpdatedActivityDetail(updatedActivity){
-        console.log('updatedActivity',updatedActivity);
+        updatedActivity.Coordinator_url = $rootScope.userDetail.data[0]._url;
+        console.log( updatedActivity.Coordinator_url);
+        console.log('updatedActivity',updateActivity);
          userInfoService.updateActivity(updatedActivity).then(function(data){
+            console.log('updatedataresult',data)
             $scope.modal.hide();
          },function(error){
             console.log(error);
@@ -122,23 +77,82 @@ angular
     }
 
     function deleteActivity(activityToDelete){
-        userInfoService.deleteActivity(activityToDelete).then(function(data){
-            console.log('deleted Succ',data);
-            showActivity();
-        },function(error){
-            console.log(error);
-        });
+        userAuthenticationService.confirm('','Do You Want To Delet Activity?','Yes','No',function(){
+            userInfoService.deleteActivity(activityToDelete).then(function(data){
+                console.log('deleted Succ',data);
+                showActivity();
+            },function(error){
+                console.log(error);
+            });
+        },null)
+        
     }
 
     function showActivity(){
         console.log('$rootScope.userDetail.data[0].SB_Region',$rootScope.userDetail.data[0].SB_Region);
         userInfoService.getUserActivities($rootScope.userDetail.data[0].SB_Region).then(function(activityData){
             vm.activityData = activityData.data;
-            console.log('region data', vm.activityData);
         },function(error){
             console.log(error);
         });
     }
 
+    function closeModel(){
+          $scope.modal.hide();
+    }  
+
+    function showUser(){
+        userInfoService.getUser().then(function(userlist){
+            vm.userlist = userlist;
+            console.log(' vm.userlist', vm.userlist);
+        },function(error){
+            console.log('Error in getting all userList',error);
+        })
+    }
+
+    function deleteUser(user){
+        userAuthenticationService.confirm('','Do You Want To Delet This User?','Yes','No',function(){
+            userInfoService.deleteActivity(user).then(function(data){
+                console.log('deleted Succ',data);
+                showUser();
+            },function(error){
+                console.log(error);
+            });
+        },null)
+        
+    }
+
+    function updateUser(userDetail){
+        vm.showUserDetail = userDetail;
+        userAuthenticationService.getProfession().then(function(userProfession){
+            vm.userProfessionList = userProfession;
+        },function(error){
+             console.log(error);
+        })
+
+        userInfoService.getUserRole().then(function(userRole){
+            vm.userRole = userRole.data;
+        },function(error){
+             console.log(error);
+        })
+        console.log('showUp',vm.showUserDetail);
+        $ionicModal.fromTemplateUrl('editUser.html', {
+            scope: $scope,
+            animation: 'slide-in-up'
+        }).then(function(modal) {
+            $scope.modal = modal;
+            $scope.modal.show();
+        });
+    }
+
+    function saveUpdatedUserDetail(updatedUserDetail){
+        console.log('updatedActivity',updatedUserDetail);
+         userInfoService.updateUserDetail(updatedUserDetail).then(function(data){
+            console.log('updatedataresult',data)
+            $scope.modal.hide();
+         },function(error){
+            console.log(error);
+         });
+    }
   }
